complete authentication, add registration page

+++ controllers/parkCtrl.js
@@ -8,17 +8,19 @@ module.exports = {
         });
     },
     park_detail: (req, res) => {
-        const {_id} = req.params;
-        Park.findOne({_id: _id}, (error, foundPark) => {
-            if(error) {
-                return error
-            } else {
-                res.render('/pages/parkUpdate', {
-                    copyrightYear: siteData.year,
-                    foundPark: foundPark
-                });
-            }
-        })
+        if(req.isAuthenticated()) {
+            const {_id} = req.params;
+            Park.findOne({_id: _id}, (error, foundPark) => {
+                if(error) {
+                    return error
+                } else {
+                    res.render('/pages/parkUpdate', {
+                        copyrightYear: siteData.year,
+                        foundPark: foundPark
+                    });
+                }
+            });
+        }
     },
 
     createPark: (req, res) => {
@@ -42,35 +44,39 @@ module.exports = {
     },
 
     park_update_put: (req, res) => {
-        const { _id } = req.params;
-        const {parkName, parkStreet, parkCity, parkState, parkZip, helpText, userName, userEmail} = req.body;
-        Park.findByIdAndUpdate(_id, {$set: {
-            parkName: parkName,
-            parkStreet: parkStreet,
-            parkCity: parkCity,
-            parkState: parkState,
-            parkZip: parkZip,
-            helpText: helpText,
-            userName: userName,
-            userEmail: userEmail
-        }}, {new: true}, error => {
-            if(error) {
-                return error;
-            } else {
-                res.redirect('/admin/parks')
-            }
-        });
+        if(req.isAuthenticated()) {
+            const { _id } = req.params;
+            const {parkName, parkStreet, parkCity, parkState, parkZip, helpText, userName, userEmail} = req.body;
+            Park.findByIdAndUpdate(_id, {$set: {
+                parkName: parkName,
+                parkStreet: parkStreet,
+                parkCity: parkCity,
+                parkState: parkState,
+                parkZip: parkZip,
+                helpText: helpText,
+                userName: userName,
+                userEmail: userEmail
+            }}, {new: true}, error => {
+                if(error) {
+                    return error;
+                } else {
+                    res.redirect('/admin/parks')
+                }
+            });
+        }
     },
 
     // this may belong in the park controller
     park_delete: (req, res) => {
-        const {_id} = req.params;
-        Park.deleteOne({_id: _id}, error => {
-            if(error) {
-                return error;
-            } else {
-                res.redirect('/admin/parks')
-            }
-        })
+        if(req.isAuthenticated()) {
+            const {_id} = req.params;
+            Park.deleteOne({_id: _id}, error => {
+                if(error) {
+                    return error;
+                } else {
+                    res.redirect('/admin/parks')
+                }
+            });
+        }
     }
 }
